+++ src/components/NicknameInput.tsx
@@ -1,5 +1,5 @@
 export default function NicknameInput(
-  props: React.ComponentPropsWithRef<"input">
+  props: React.ComponentPropsWithRef<"input">,
 ) {
   return (
     <div className="flex-col-center px-8">
